Pass matrix projection matrices to shader
