+++ src/main.rs
@@ -3,6 +3,7 @@ use render::RenderSystem;
 mod critter;
 mod data;
 mod game;
+mod graphics;
 mod render;
 
 pub fn main() {
