+++ src/render/window.rs
@@ -1,14 +1,6 @@
 use wgpu::winit::{Event, EventsLoop, Window, WindowEvent};
 use winit::dpi::LogicalSize;
 
-#[allow(dead_code)]
-pub fn cast_slice<T>(data: &[T]) -> &[u8] {
-  use std::mem::size_of;
-  use std::slice::from_raw_parts;
-
-  unsafe { from_raw_parts(data.as_ptr() as *const u8, data.len() * size_of::<T>()) }
-}
-
 #[derive(PartialEq, Eq)]
 pub enum WindowStatus {
   Open,
